# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "KR/US Watchlist AI Stock Analysis API\n\n## Modules\n- Stock analysis: trigger AI analysis\n- History: query historical analysis reports\n- Stock data: fetch market quotes\n\n## Authentication\nNo authentication required in the current version",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analysis/analyze": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "종목 분석 실행",
        "description": "AI 종목 분석 작업을 시작합니다. 동기/비동기 모드를 지원하며 비동기 모드에서는 중복 제출을 막습니다.",
        "operationId": "trigger_analysis_api_v1_analysis_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "분석 완료(동기 모드)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultResponse"
                }
              }
            }
          },
          "202": {
            "description": "분석 작업 접수됨(비동기 모드)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "400": {
            "description": "요청 파라미터 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "해당 종목 분석이 진행 중이라 중복 제출이 거부됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateTaskErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "분석 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/tasks": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "분석 작업 목록 조회",
        "description": "현재 분석 작업을 조회합니다. 상태로 필터링할 수 있습니다.",
        "operationId": "get_task_list_api_v1_analysis_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "필터 상태: pending, processing, completed, failed. 여러 값은 쉼표로 구분합니다.",
              "title": "Status"
            },
            "description": "필터 상태: pending, processing, completed, failed. 여러 값은 쉼표로 구분합니다."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "반환 개수 제한",
              "default": 20,
              "title": "Limit"
            },
            "description": "반환 개수 제한"
          }
        ],
        "responses": {
          "200": {
            "description": "작업 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis/status/{task_id}": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "분석 작업 상태 조회",
        "description": "task_id로 단일 작업 상태를 조회합니다.",
        "operationId": "get_analysis_status_api_v1_analysis_status__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "작업 상태",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                }
              }
            }
          },
          "404": {
            "description": "작업을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisResultResponse": {
        "properties": {
          "query_id": {
            "type": "string",
            "title": "Query Id",
            "description": "분석 기록 고유 ID"
          },
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "종목 코드"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name",
            "description": "종목명"
          },
          "report": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Report",
            "description": "분석 리포트"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "생성 시간"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "stock_code",
          "created_at"
        ],
        "title": "AnalysisResultResponse",
        "description": "Analysis result response model.",
        "example": {
          "created_at": "2024-01-01T12:00:00",
          "query_id": "abc123def456",
          "report": {
            "summary": {
              "operation_advice": "보유",
              "sentiment_score": 75
            }
          },
          "stock_code": "005930",
          "stock_name": "삼성전자"
        }
      },
      "AnalyzeRequest": {
        "properties": {
          "stock_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Code",
            "description": "단일 종목 코드",
            "example": "005930"
          },
          "stock_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Codes",
            "description": "여러 종목 코드(stock_code와 둘 중 하나 사용)",
            "example": [
              "005930",
              "AAPL"
            ]
          },
          "report_type": {
            "type": "string",
            "pattern": "^(simple|detailed)$",
            "title": "Report Type",
            "description": "리포트 유형",
            "default": "detailed"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "캐시를 무시하고 강제로 새로고침할지 여부",
            "default": true
          },
          "async_mode": {
            "type": "boolean",
            "title": "Async Mode",
            "description": "비동기 모드 사용 여부",
            "default": false
          }
        },
        "type": "object",
        "title": "AnalyzeRequest",
        "description": "Analysis request model.",
        "example": {
          "async_mode": false,
          "force_refresh": false,
          "report_type": "detailed",
          "stock_code": "005930"
        }
      },
...
      "DuplicateTaskErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "오류 유형",
            "default": "duplicate_task"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "오류 메시지"
          },
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "종목 코드"
          },
          "existing_task_id": {
            "type": "string",
            "title": "Existing Task Id",
            "description": "이미 존재하는 작업 ID"
          }
        },
        "type": "object",
        "required": [
          "message",
          "stock_code",
          "existing_task_id"
        ],
        "title": "DuplicateTaskErrorResponse",
        "description": "Duplicate task error response model.",
        "example": {
          "error": "duplicate_task",
          "existing_task_id": "abc123def456",
          "message": "종목 005930 분석 작업이 이미 진행 중입니다",
          "stock_code": "005930"
        }
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error detail",
            "example": "요청 파라미터 오류"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error information"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Error response.",
        "example": {
          "error": "not_found",
          "message": "리소스를 찾을 수 없습니다"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskAccepted": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "상태 조회에 사용하는 작업 ID"
          },
          "status": {
            "type": "string",
            "pattern": "^(pending|processing)$",
            "title": "Status",
            "description": "작업 상태"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "안내 메시지"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskAccepted",
        "description": "Asynchronous task acceptance response.",
        "example": {
          "message": "Analysis task accepted",
          "status": "pending",
          "task_id": "task_abc123"
        }
      },
...
      "TaskListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "전체 작업 수"
          },
          "pending": {
            "type": "integer",
            "title": "Pending",
            "description": "대기 중인 작업 수"
          },
          "processing": {
            "type": "integer",
            "title": "Processing",
            "description": "처리 중인 작업 수"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskInfo"
            },
            "type": "array",
            "title": "Tasks",
            "description": "작업 목록"
          }
        },
        "type": "object",
        "required": [
          "total",
          "pending",
          "processing",
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "Task list response model.",
        "example": {
          "pending": 1,
          "processing": 2,
          "tasks": [],
          "total": 3
        }
      },
      "TaskStatus": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "작업 ID"
          },
          "status": {
            "type": "string",
            "pattern": "^(pending|processing|completed|failed)$",
            "title": "Status",
            "description": "작업 상태"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress",
            "description": "진행률 (0-100)"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResultResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "분석 결과(completed 상태에서만 존재)"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "오류 메시지(failed 상태에서만 존재)"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskStatus",
        "description": "Task status model.",
        "example": {
          "progress": 100,
          "status": "completed",
          "task_id": "task_abc123"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}